Continued to play with styling and exact paths to try to reformat Header
Co-authored by: jeffcoat.megan@gmail.com
