+++ client/src/auth/Login.jsx
@@ -2,6 +2,8 @@ import React, {useState} from "react"
 
 import axios from "axios"
 
+import Header from '../components/Header'
+
 function Login(props) {
     const [creds, setCreds] = useState({});
