+++ client/src/auth/Register.jsx
@@ -2,35 +2,6 @@ import React, { useState } from "react";
 import axios from "axios";
 import styled from 'styled-components';
 
-const FormContainer = styled.section`
-    margin: 0 auto;
-    padding: 7.5%;
-    border-radius: 5px;
-    border: 2px solid red;
-`;
-
-const Signup = styled.form`
-    display: flex;
-    flex-direction: column;
-    margin: 8%;
-
-    input {
-        width: 100%;
-        padding: 5% 5px;
-        margin: 5% ;
-        border: 1px solid red;
-        border-radius: 5px;
-    }
-
-    button {
-        width: 110%;
-        padding: 5% 5px;
-        margin: 5%;
-        border: 1px solid red;
-        border-radius: 5px; 
-    }
-`;
-
 function Register(props) {
     const [creds, setCreds] = useState({});
   
@@ -55,7 +26,9 @@ function Register(props) {
     };
   
     return (
+      
       <div className="Login-Page">
+        
         <form className="login-form" onSubmit={handleLogin}>
           <h1 className="Login-Title">Register</h1>
           <input
@@ -81,7 +54,9 @@ function Register(props) {
           />
           <button className="registerBtn">Register</button>
         </form>
+        
       </div>
+      
     );
   }
